% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FarPavaSourceFunctions.R
\name{Fun_cdf_Pred}
\alias{Fun_cdf_Pred}
\title{Monotone Functional Auto-Regression Predictions}
\usage{
Fun_cdf_Pred(x, y, xnew, p = 1, ahead, ci = 0.95)
}
\arguments{
\item{x}{Observed locations of function}

\item{y}{Function values at observed locations through time, row is location, column is time.}

\item{xnew}{Evaluate locations of function. When this is same as 'x', the return value is same as vector auto regression.}

\item{p}{Integer for the lag order (default is p=1)}

\item{ahead}{An integer specifying the number of forecast steps}

\item{ci}{The forecast confidence interval (default is 0.95)}
}
\value{
forecast value of the function for each step at evaluate locations
}
\description{
Monotone Functional Auto-Regression Predictions
}
