% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FarPavaSourceFunctions.R
\name{iso_reg}
\alias{iso_reg}
\title{Isotonic (Increasing) Regression}
\usage{
iso_reg(a)
}
\arguments{
\item{a}{A numeric vector}
}
\value{
A vector with increasing terms using Pool-Adjacent-Violators Algorithm
}
\description{
Isotonic (Increasing) Regression
}
\examples{
iso_reg(c(1,5,2,3,1))


}
