% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FarPavaSourceFunctions.R
\name{mchi}
\alias{mchi}
\title{Monotone Cubic Hermite Interpolation}
\usage{
mchi(x, y, xnew)
}
\arguments{
\item{x}{Observed locations on a function}

\item{y}{Estimated Value of the function at each observed locations}

\item{xnew}{Evaluate locations}
}
\value{
Estimated Value of the function at each evaluate locations
}
\description{
Monotone Cubic Hermite Interpolation
}
\examples{
mchi(c(1,2,3),c(0.6,0.7,0.77),c(4,5))
}
