% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FarPavaSourceFunctions.R
\name{RMSE}
\alias{RMSE}
\title{Title}
\usage{
RMSE(a, b)
}
\arguments{
\item{a}{A numeric vector}

\item{b}{A numeric vector that has same length as $a$}
}
\value{
The rooted mean squared distance between $a$ and $b$
}
\description{
Title
}
