% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FarPavaSourceFunctions.R
\name{tru_func}
\alias{tru_func}
\title{Truncate vector between 0 and 1}
\usage{
tru_func(an)
}
\arguments{
\item{an}{A numeric array}
}
\value{
The array with every entry bigger than 1 shrinked into 1 and every entry smaller than 0 shrinked into 0
}
\description{
Truncate vector between 0 and 1
}
\examples{
tru_func(c(0,2,1,-2))
}
