% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FarPavaSourceFunctions.R
\name{Var_Pred}
\alias{Var_Pred}
\title{Vector Auto-Regression Predictions}
\usage{
Var_Pred(y, p = 1, ahead, ci = 0.95)
}
\arguments{
\item{y}{A data matrix with each column is values of function on observed locations}

\item{p}{Integer for the lag order (default is p=1)}

\item{ahead}{An integer specifying the number of forecast steps.}

\item{ci}{The forecast confidence interval (default is 0.95)}
}
\value{
A list with 3 items: "forecast" forecast value of the function for each step at locations; "lower" lower confidence band; "upper" upper confidence band
}
\description{
Vector Auto-Regression Predictions
}
\examples{
y_ram = matrix(runif(48),nrow=3)
Var_Pred(y_ram, p=1, ahead=5, ci=0.95)
}
